Take methods from SimplePostTool and move them to BinaryUtils
